Chromium patch fixing macOS shared textures with skia
This is a patch corresponding to CL [1] authored by Carol Wolfking.
Currently SkiaOutputSurface will create a characteristic that
backend_format will be kRectangle if the build is for macOS. But when
the output_device is SkiaOutputDeviceOffscreen, the surface check will
fail at GrSurfaceProxy::validateSurface saying:
surface->asTexture()->textureType() != this->textureType()

This fix make SkiaOutputDeviceOffscreen use same backend_format when
creating Backbuffers.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5266558